requirements.txt
# requirements.txt - Cross-platform compatibility
# Core packages
albumentations>=1.3.0
opencv-python>=4.6.0
numpy>=1.22.0
PyYAML>=6.0
ultralytics>=8.0.0
matplotlib>=3.7.0
Pillow>=9.4.0
psutil>=5.9.0
requests>=2.28.0
optuna>=3.0.0
joblib>=1.2.0

# PyTorch - platform-specific installations should be done separately
# Use the appropriate command from pytorch.org for your system
# For example: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# Placeholder entries:
torch>=2.0.0; platform_system != "Linux" or platform_system != "Windows"
torchvision>=0.15.0; platform_system != "Linux" or platform_system != "Windows"

# Platform-specific dependencies
wmi>=1.5.1; platform_system == "Windows"

# Linux AMD GPU support packages (when ROCm is needed)
# Note: ROCm PyTorch should be installed separately using AMD's instructions
# https://rocm.docs.amd.com/en/latest/deploy/linux/quick_start.html

# Testing dependencies
pytest>=7.3.1
pytest-cov>=4.1.0
flake8>=6.0.0
mypy>=1.3.0
